Added test for admin view, all tests passing
